backend/requirements.txt
# Flask and web framework dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Machine Learning and Data Science
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.21.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
Pillow>=10.0.0
imageio>=2.30.0

# Data handling
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.60.0

# Development and testing (optional)
pytest>=7.0.0
pytest-flask>=1.2.0
